requirements.txt
absl-py==0.5.0
astor==0.7.1
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
Keras==2.2.2
Keras-Applications==1.0.5
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.2
opencv-python==3.4.3.18
Pillow==5.3.0
protobuf==3.6.1
psycopg2==2.7.5
PyQt5==5.11.2
PyQt5-sip==4.19.12
PythonVideoDevices==1.0
PyYAML==3.13
scipy==1.1.0
sip==4.19.8
six==1.11.0
SQLAlchemy==1.2.12
tensorboard==1.11.0
tensorflow-gpu==1.11.0
termcolor==1.1.0
webcolors==1.8.1
Werkzeug==0.14.1
WindowsDevices==1.0
zope.interface==4.5.0
